Shipping price can be float not int. Checkout page should show tax including shipping price, but basket not.

+++ system/modules/isotope/config/database.sql
@@ -387,7 +387,7 @@ CREATE TABLE `tl_shipping_modules` (
   `countries` blob NULL,
   `minimum_total` int(10) NOT NULL default '0',
   `maximum_total` int(10) NOT NULL default '0',
-  `price` int(10) NOT NULL default '0',
+  `price` float NOT NULL default '0',
   `flatCalculation` varchar(10) NOT NULL default '',
   `enabled` char(1) NOT NULL default '',
   PRIMARY KEY  (`id`),
